refactor #47 패키지 구조 변경으로 인한 수정
- controller, domain, dto, repository, service 외 패키지를 common 패키지로 옮김
